+++ packages/theme/src/composables/config/index.ts
@@ -106,11 +106,16 @@ export namespace Theme {
     nextText?: string
     /**
      * 是否展示当前正在浏览的文章在左侧
-     * @default false
+     * @default true
      */
     showSelf?: boolean
     filter?: (page: Theme.PageData) => boolean
     empty?: string | boolean
+    /**
+     * 设置推荐文章的展示风格
+     * @default 'card'
+     */
+    style?: 'card' | 'sidebar'
   }
 
   export interface HomeBlog {
@@ -264,6 +269,11 @@ export namespace Theme {
      * @default false
      */
     mermaid?: any
+    /**
+     * 设置解析 frontmatter 里 date 的时区
+     * @default 8 => 'UTC+8'
+     * */
+    timeZone?: number
   }
 
   export interface Config extends DefaultTheme.Config {