+++ packages/blogpress/CHANGELOG.md
@@ -1,4 +1,14 @@
+---
+publish: false
+---
 # blogpress
+
+## 2.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @sugarat/theme@0.1.39
 
 ## 2.0.3