+++ packages/theme/docs/changelog.md
@@ -14,6 +14,17 @@ publish: true
 ```sh
 pnpm up @sugarat/theme
 ```
+
+## 0.1.39 (2023/08/05)
+
+### Patch Changes
+
+- feat: 支持配置解析 `frontmatter` 的时区
+- fix: 单独使用主题的define方法报错
+- feat: 支持切换主题 sidebar 风格
+- fix: 相关文章的序号又从 1 开始
+- fix: 2位数序号展示折行
+
 ## 0.1.38 (2023/07/29)
 
 ### Patch Changes